# contracts/openapi/p-api.yaml
openapi: 3.1.0
info:
  title: DX-LAB Process Core (P) API
  version: 0.1.0
  description: >-
    Hợp đồng giao tiếp RESTful API chuẩn của dịch vụ lõi P cho các tác vụ
    tiếp nhận, điều phối và tra cứu quy trình nghiệp vụ số DX-OS.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:3000
    description: Cục bộ (Local Development)
  - url: http://p-process:3000
    description: Mạng nội bộ Docker Compose (application network)
paths:
...
  /api/v1/resources:
    get:
      summary: Tra cứu danh sách SOP và FAQ đang có hiệu lực
      operationId: listResources
      security:
        - keycloakBearer: [portal:read]
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [sop, faq]
          description: Lọc theo loại tài liệu
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Từ khóa tìm kiếm theo tiêu đề, mã, tóm tắt hoặc nội dung
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Danh sách tài liệu tri thức đang có hiệu lực
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - total
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResourceListItem'
                  total:
                    type: integer
        '400':
          description: Tham số tìm kiếm hoặc phân trang không hợp lệ
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '401':
          description: Thiếu hoặc sai bearer token
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '403':
          description: Thiếu scope portal:read hoặc tài khoản không có quyền
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '503':
          description: Dịch vụ tạm thời không khả dụng
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
  /api/v1/resources/ai/retrieve:
    get:
      summary: Truy xuất tri thức SOP và FAQ đã công bố phục vụ AI retrieval
      operationId: retrieveAiResources
      security:
        - keycloakBearer: [portal:read]
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: Câu truy vấn cần tra cứu tri thức
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 20
      responses:
        '200':
          description: Kết quả truy xuất tri thức đã công bố kèm metadata phiên bản và nguồn gốc
          content:
            application/json:
              schema:
                type: object
                required:
                  - query
                  - results
                properties:
                  query:
                    type: string
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/AiResourceItem'
        '400':
          description: Thiếu tham số query
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '401':
          description: Thiếu hoặc sai bearer token
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '403':
          description: Thiếu scope portal:read hoặc tài khoản không có quyền
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '503':
          description: Dịch vụ tạm thời không khả dụng
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
  /api/v1/resources/{id}:
    get:
      summary: Tra cứu chi tiết tài liệu SOP/FAQ đã công bố theo ID
      operationId: getResourceById
      security:
        - keycloakBearer: [portal:read]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID (UUID) hoặc mã định danh của tài liệu
      responses:
        '200':
          description: Chi tiết toàn văn tài liệu đã công bố
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceDetail'
        '401':
          description: Thiếu hoặc sai bearer token
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '403':
          description: Thiếu scope portal:read hoặc tài liệu không được phép truy cập
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '404':
          description: Không tìm thấy tài liệu hoặc tài liệu chưa công bố
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '503':
          description: Dịch vụ tạm thời không khả dụng
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
components:
  securitySchemes:
    keycloakBearer:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /realms/dxlab/protocol/openid-connect/auth
          tokenUrl: /realms/dxlab/protocol/openid-connect/token
          scopes:
            portal:read: Đọc định danh và membership Portal hiện hành
            tickets:read: Đọc ticket theo phạm vi trách nhiệm
            tickets:download: Tải tệp của ticket được phép
            tickets:write: Xử lý ticket do mình phụ trách trong nhóm
  schemas:
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Trường tiêu đề không được để trống
        instance:
          type: string
          example: /api/v1/tickets
        code:
          type: string
          example: validation_failed
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    ResourceListItem:
      type: object
      additionalProperties: false
      required:
        - id
        - code
        - type
        - title
        - version
        - effectiveDate
        - status
        - statusLabel
        - approverName
        - publishedAt
        - summary
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: SOP-TKT-001
        type:
          type: string
          enum: [sop, faq]
        title:
          type: string
          example: Quy trình tiếp nhận và phân công ticket
        version:
          type: string
          example: 1.0.0
        effectiveDate:
          type: string
          example: '2026-01-01'
        status:
          type: string
          enum: [published]
        statusLabel:
          type: string
          example: Đang có hiệu lực
        approverName:
          type: string
          example: Ban Giám đốc
        publishedAt:
          type: string
          format: date-time
        summary:
          type: string
    ResourceDetail:
      type: object
      additionalProperties: false
      required:
        - id
        - code
        - type
        - title
        - version
        - effectiveDate
        - status
        - statusLabel
        - approverName
        - publishedAt
        - summary
        - content
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: SOP-TKT-001
        type:
          type: string
          enum: [sop, faq]
        title:
          type: string
        version:
          type: string
          example: 1.0.0
        effectiveDate:
          type: string
          example: '2026-01-01'
        status:
          type: string
          enum: [published]
        statusLabel:
          type: string
          example: Đang có hiệu lực
        approverName:
          type: string
        publishedAt:
          type: string
          format: date-time
        summary:
          type: string
        content:
          type: string
    AiResourceItem:
      type: object
      additionalProperties: false
      required:
        - id
        - code
        - type
        - title
        - version
        - effectiveDate
        - approverName
        - publishedAt
        - summary
        - content
        - provenance
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        type:
          type: string
          enum: [sop, faq]
        title:
          type: string
        version:
          type: string
        effectiveDate:
          type: string
        approverName:
          type: string
        publishedAt:
          type: string
          format: date-time
        summary:
          type: string
        content:
          type: string
        provenance:
          type: object
          additionalProperties: false
          required:
            - source
            - docCode
            - version
            - publishedAt
          properties:
            source:
              type: string
              example: dx_core.resources
            docCode:
              type: string
              example: SOP-TKT-001
            version:
              type: string
              example: 1.0.0
            publishedAt:
              type: string
              format: date-time